+++ test/postgrest/embed_inner_join_test.go
@@ -204,16 +204,11 @@ func TestPostgREST_EmbedInnerJoin(t *testing.T) {
 			ExpectedHeaders: map[string]string{"Content-Range": "0-7/8"},
 			Status:          200,
 		},
-		// 	it "works with views" $ do
-		// 	  get "/books?select=title,authors!inner(name)&authors.name=eq.George%20Orwell" `shouldRespondWith`
-		// 		[json| [{"title":"1984","authors":{"name":"George Orwell"}}] |]
-		// 		{ matchHeaders = [matchContentTypeJson] }
-		// 	  request methodHead "/books?select=title,authors!inner(name)&authors.name=eq.George%20Orwell" [("Prefer", "count=exact")] mempty
-		// 		`shouldRespondWith` ""
-		// 		{ matchStatus  = 200
-		// 		, matchHeaders = [ matchContentTypeJson
-		// 						 , "Content-Range" <:> "0-0/1" ]
-		// 		}
+		{
+			Description: "m2o inner join works with views",
+			Query:       "/books?select=title,authors!inner(name)&authors.name=eq.George%20Orwell",
+			Expected:    `[{"title":"1984","authors":{"name":"George Orwell"}}]`,
+		},
 
 		//   context "one-to-many relationships" $ do
 		// 	it "ignores empty array embeddings while the default left join doesn't" $ do
@@ -428,16 +423,11 @@ func TestPostgREST_EmbedInnerJoin(t *testing.T) {
 			ExpectedHeaders: map[string]string{"Content-Range": "0-1/2"},
 			Status:          200,
 		},
-		// 	it "works with views" $ do
-		// 	  get "/authors?select=*,books!inner(*)&books.title=eq.1984" `shouldRespondWith`
-		// 		[json| [{"id":1,"name":"George Orwell","books":[{"id":1,"title":"1984","publication_year":1949,"author_id":1,"first_publisher_id":1}]}] |]
-		// 		{ matchHeaders = [matchContentTypeJson] }
-		// 	  request methodHead "/authors?select=*,books!inner(*)&books.title=eq.1984" [("Prefer", "count=exact")] mempty
-		// 		`shouldRespondWith` ""
-		// 		{ matchStatus  = 200
-		// 		, matchHeaders = [ matchContentTypeJson
-		// 						 , "Content-Range" <:> "0-0/1" ]
-		// 		}
+		{
+			Description: "o2m inner join works with views",
+			Query:       "/authors?select=*,books!inner(*)&books.title=eq.1984",
+			Expected:    `[{"id":1,"name":"George Orwell","books":[{"id":1,"title":"1984","publication_year":1949,"author_id":1,"first_publisher_id":1}]}]`,
+		},
 
 		//   context "many-to-many relationships" $ do
 		// 	it "ignores empty array embeddings while the default left join doesn't" $ do
@@ -592,19 +582,16 @@ func TestPostgREST_EmbedInnerJoin(t *testing.T) {
 			ExpectedHeaders: map[string]string{"Content-Range": "0-1/2"},
 			Status:          200,
 		},
-		// 	it "works with views" $ do
-		// 	  get "/actors?select=*,films!inner(*)&films.title=eq.douze%20commandements" `shouldRespondWith`
-		// 		[json| [{"id":1,"name":"john","films":[{"id":12,"title":"douze commandements"}]}] |]
-		// 		{ matchHeaders = [matchContentTypeJson] }
-		// 	  get "/films?select=*,actors!inner(*)&actors.name=eq.john" `shouldRespondWith`
-		// 		[json| [{"id":12,"title":"douze commandements","actors":[{"id":1,"name":"john"}]}] |]
-		// 		{ matchHeaders = [matchContentTypeJson] }
-		// 	  request methodHead "/actors?select=*,films!inner(*)&films.title=eq.douze%20commandements" [("Prefer", "count=exact")] mempty
-		// 		`shouldRespondWith` ""
-		// 		{ matchStatus  = 200
-		// 		, matchHeaders = [ matchContentTypeJson
-		// 						 , "Content-Range" <:> "0-0/1" ]
-		// 		}
+		{
+			Description: "m2m inner join works with views (actors→films)",
+			Query:       "/actors?select=*,films!inner(*)&films.title=eq.douze%20commandements",
+			Expected:    `[{"id":1,"name":"john","films":[{"id":12,"title":"douze commandements"}]}]`,
+		},
+		{
+			Description: "m2m inner join works with views (films→actors)",
+			Query:       "/films?select=*,actors!inner(*)&actors.name=eq.john",
+			Expected:    `[{"id":12,"title":"douze commandements","actors":[{"id":1,"name":"john"}]}]`,
+		},
 
 		//   it "works with m2o and m2m relationships combined" $ do
 		// 	get "/projects?select=name,clients!inner(name),users!inner(name)" `shouldRespondWith`
@@ -748,17 +735,14 @@ func TestPostgREST_EmbedInnerJoin(t *testing.T) {
 		// 		{"id":9,"authors":{"name":"Ken Kesey"},"publishers":{"name":"Viking Press & Signet Books"}}] |]
 		// 	  { matchHeaders = [matchContentTypeJson] }
 
-		//@@ views
-
-		// {
-		// 	Description: "works alongside another embedding",
-		// 	Query:       "/books?select=id,authors(name),publishers!inner(name)&id=gte.7",
-		// 	Expected: `[
-		// 				{"id":7,"authors":{"name":"Harper Lee"},"publishers":{"name":"J. B. Lippincott & Co."}},
-		// 				{"id":8,"authors":{"name":"Kurt Vonnegut"},"publishers":{"name":"Delacorte"}},
-		// 				{"id":9,"authors":{"name":"Ken Kesey"},"publishers":{"name":"Viking Press & Signet Books"}}]`,
-		// 	Headers: nil,
-		// },
+		{
+			Description: "works alongside another embedding (views)",
+			Query:       "/books?select=id,authors(name),publishers!inner(name)&id=gte.7",
+			Expected: `[
+						{"id":7,"authors":{"name":"Harper Lee"},"publishers":{"name":"J. B. Lippincott & Co."}},
+						{"id":8,"authors":{"name":"Kurt Vonnegut"},"publishers":{"name":"Delacorte"}},
+						{"id":9,"authors":{"name":"Ken Kesey"},"publishers":{"name":"Viking Press & Signet Books"}}]`,
+		},
 		// 	request methodHead "/books?select=id,authors(name),publishers!inner(name)&id=gte.7" [("Prefer", "count=exact")] mempty
 		// 	  `shouldRespondWith` ""
 		// 	  { matchStatus  = 200